requirements.txt
requests
bs4
flask
lxml
pandas
mysql-connector-python
pymysql